apps/workspaces/serializers.py
from rest_framework import serializers
from .models import Workspace

class WorkspaceSerializer(serializers.ModelSerializer):
    class Meta:
        model = Workspace
        fields = ['id', 'name', 'join_code', 'created_at']
        read_only_fields = ['join_code']
